frontend/detector/requirements.txt
# Requirements for Network Security Detector
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
